requirements.txt
streamlit==1.28.1
chromadb==0.5.3
sentence-transformers==2.7.0
pypdf2==3.0.1
python-pptx==0.6.22
pandas==1.5.3  # Changed to a version compatible with Python 3.10
python-docx==0.8.11
markdown==3.5.1
google-generativeai==0.7.1
numpy==1.24.3
uuid==1.30
